pyproject.toml
[project]
name = "01-github-backlog-issues"
version = "0.1.0"
description = "GitPulse: GitHub Issues Medallion Pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.32",
    "python-dotenv>=1.0",
    "tenacity>=9.0",
    "duckdb>=1.1",
    "pyarrow>=16.0",
    "boto3>=1.34",
    "pytz>=2026.1.post1",
]

[dependency-groups]
dev = ["pytest>=8.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = ["integration: requires live MinIO and pipeline data (deselected by default)"]
addopts = "-m 'not integration'"
